Created separate controller
